Create tetrous.py
Adding the tetrous.py file, which is spokes an hero gun icon code. Since I had to download a new gun image to save as 'exit.png' to run it, it is different but if spoke uploads the gun icon she used from her computer it could be merged in and replace the image I used in this commit.